Вынес логику похода на сервер во ВьюМодель
