+++ app/src/main/java/com/aroman/gitandroid/App.kt
@@ -2,8 +2,10 @@ package com.aroman.gitandroid
 
 import android.app.Application
 import android.content.Context
+import com.aroman.gitandroid.data.GitRetrofitImpl
 
 class App : Application() {
+    val retrofitImpl by lazy { GitRetrofitImpl() }
 }
 
 val Context.app: App
