+++ app/src/main/java/com/aroman/gitandroid/ui/GitAndroidContract.kt
@@ -1,0 +1,14 @@
+package com.aroman.gitandroid.ui
+
+import androidx.annotation.MainThread
+import com.aroman.gitandroid.domain.entities.GitServerResponseData
+import com.aroman.gitandroid.utils.Publisher
+
+class GitAndroidContract {
+    interface UserDetailsViewModel {
+        val data: Publisher<GitServerResponseData>
+
+        @MainThread
+        fun getUser(login: String)
+    }
+}
