+++ app/src/main/java/com/aroman/gitandroid/ui/UserDetailsFragment.kt
@@ -1,29 +1,28 @@
 package com.aroman.gitandroid.ui
 
 import android.os.Bundle
+import android.os.Handler
+import android.os.Looper
 import android.text.Html
 import android.text.method.LinkMovementMethod
-import android.util.Log
 import android.view.Gravity
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
-import com.aroman.gitandroid.R
 import com.aroman.gitandroid.data.GitRetrofitImpl
 import com.aroman.gitandroid.databinding.FragmentUserDetailsBinding
 import com.aroman.gitandroid.domain.entities.GitServerResponseData
 import com.squareup.picasso.Picasso
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
 
 private const val LOGIN = "login"
 
 class UserDetailsFragment() : Fragment() {
     private var login = "login"
     private lateinit var binding: FragmentUserDetailsBinding
+    private val viewModel: UserDetailsViewModel by lazy { UserDetailsViewModel(GitRetrofitImpl()) }
+    private val handler: Handler by lazy { Handler(Looper.getMainLooper()) }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -44,32 +43,18 @@ class UserDetailsFragment() : Fragment() {
         super.onViewCreated(view, savedInstanceState)
         binding.loginTextView.text = login
 
-        GitRetrofitImpl().getGitRetrofitImpl().listRepos(login).enqueue(
-            object : Callback<List<GitServerResponseData>> {
-                override fun onResponse(
-                    call: Call<List<GitServerResponseData>>,
-                    response: Response<List<GitServerResponseData>>
-                ) {
-                    if (response.isSuccessful && !response.body().isNullOrEmpty()) {
-                        for (repo in response.body()!!) {
-                            Picasso.get().load(repo.owner.avatarUrl).into(binding.avatarImageView)
-                            initRepoList(repo)
-                        }
-                    }
-                }
-
-                override fun onFailure(call: Call<List<GitServerResponseData>>, t: Throwable) {
-                    Log.d("@@@", "onFailure: ${t.message}")
-                }
-            }
-        )
+        viewModel.data.subscribe(handler) { repo ->
+            Picasso.get().load(repo?.owner?.avatarUrl).into(binding.avatarImageView)
+            initRepoList(repo)
+        }
+        viewModel.getUser(login)
     }
 
-    private fun initRepoList(repo: GitServerResponseData) {
+    private fun initRepoList(repo: GitServerResponseData?) {
         binding.llRepoList.addView(
             TextView(requireActivity()).apply {
                 text = Html.fromHtml(
-                    "<a href=\"" + repo.repoHtmlUrl + "\">" + repo.repoName + "</a>",
+                    "<a href=\"" + repo?.repoHtmlUrl + "\">" + repo?.repoName + "</a>",
                     Html.FROM_HTML_MODE_LEGACY
                 )
                 textSize = 30f
